Remove Bundler lockfiles, ignore them in future.
This should fix #56.  I think our Travis-CI was breaking on each new
release of the gem because we suddenly stopped satisfying the version
in the included lock files.

Anyway, we shouldn't ever have lock files for Travis testing since
that defeats the purpose of CI, right?  We *want* to test against
the latest and greatest...