Implement new node 'tags' attribute behavior.
Fixes #49 (and #50) by supporting both legacy behavior, in which we
expect that Chef environments, roles and run lists are passed into
the RunDeck resource model via the "tags" attribute ... and the new
behavior, in which we expect that roles and run lists will be
passed in as custom attributes.  This gives everybody a chance to
upgrade their filters before their jobs mysteriously stop running on
the expected range of hosts!